Data parser tool add
