+++ VisionAnalysis/VisionAnalysis.csproj
@@ -122,7 +122,7 @@
     </ApplicationDefinition>
     <Compile Include="UIs\Tools\Teps\TepCanny\TepCanny.cs" />
     <Compile Include="UIs\Tools\Teps\TepCvtColor\TepCvtColor.cs" />
-    <Compile Include="UIs\Tools\Trasfer\TepDataTransfer\TepObjPaser.cs" />
+    <Compile Include="UIs\Tools\Trasfer\TepDataTransfer\TepObjParser.cs" />
     <Compile Include="UIs\Tools\Teps\TepDilate\TepDilate.cs" />
     <Compile Include="UIs\Tools\Teps\TepErode\TepErode.cs" />
     <Compile Include="UIs\Tools\Teps\TepFindContours\TepFindContours.cs" />
